[ocean] Sync repos
Summary:
Syncing repos.

fbshipit-source-id: e8d3798aa5d15bcdde4b9e0f497cac558fa989e7

+++ impl/ocean/devices/arcore/ARSessionManager.cpp
@@ -998,7 +998,7 @@ Frame ARSessionManager::extractImage(ArSession* arSession, ArFrame* arFrame, Med
 				sourcePixelStrides[planeIndex] = (unsigned int)(planePixelStride);
 
 				sourcePaddingElements[planeIndex] = (unsigned int)(planeDataRowStride - planeWidth);
-				ocean_assert(sourcePaddingElements[planeIndex] < planeDataRowStride);
+				ocean_assert(sourcePaddingElements[planeIndex] < (unsigned int)(planeDataRowStride));
 			}
 
 			int64_t timestampNs = NumericT<int64_t>::minValue();
